+++ server/routes/data/appointments.ts
@@ -1,0 +1,41 @@
+export const pastAppointments = [
+  {
+    date: 'Thursday 20 Jan 2025',
+    time: '10:00am',
+    title: 'Probation appointment',
+    location: '123 Garden Street, London SE1 7TH',
+    contact: 'Karen Smith',
+    contactLink: '/pop/new-message',
+  },
+  {
+    date: 'Thursday 15 Jan 2025',
+    time: '2:00pm',
+    title: 'Probation appointment',
+    location: 'National Probation Service, 235 Greenwich High Road, London SE10 8NB',
+    contact: 'Julie Myers',
+    contactLink: '',
+  },
+]
+
+export const upcomingAppointments = [
+  {
+    date: 'Tuesday 18 March 2025',
+    time: '10:00am',
+    title: 'Community Garden Maintenance',
+    location: '123 Garden Street, London SE1 7TH',
+    contact: 'Karen Smith',
+    contactLink: '',
+    description:
+      'Group session focused on weeding, planting seasonal vegetables, and general garden maintenance. Bring appropriate clothing for outdoor work, tools will be provided. Break times will be scheduled during the session.',
+  },
+  {
+    date: 'Thursday 21 March 2025',
+    time: '2:00pm',
+    title: 'Probation appointment',
+    location: 'National Probation Service, 235 Greenwich High Road, London SE10 8NB',
+    contact: 'Julie Myers',
+    contactLink: '',
+    description:
+      'Regular probation check-in to discuss progress, any issues faced, and upcoming milestones. Make sure to bring any requested documents or updates for the probation officer.',
+  },
+]
