+++ server/views/pages/pop/messageThread.njk
@@ -1,10 +1,9 @@
 {% extends "../../partials/layout.njk" %}
 
-{%- from "moj/components/ticket-panel/macro.njk" import mojTicketPanel -%}
 {%- from "govuk/components/button/macro.njk" import govukButton %}
-{% from "govuk/components/tag/macro.njk" import govukTag %}
+{%- from "moj/components/messages/macro.njk" import mojMessages -%}
 
-{% set pageTitle = applicationName + " - Create new message" %}
+{% set pageTitle = applicationName + " - Messages" %}
 {% set mainClasses = "app-container govuk-body" %}
 
 {% block content %}
@@ -15,92 +14,41 @@
       <a class="govuk-link" href="/pop/messages">&lt; Back to messages</a>
       <div class="govuk-grid-row govuk-!-margin-top-6 govuk-!-paddin-bottom-4">
           <h1 class="govuk-heading-l govuk-!-margin-0 govuk-!-padding-0 govuk-!-padding-bottom-6 ">{{message.subject}}</h1>
-     {{ mojTicketPanel({
-  attributes: {
-    "aria-label": "Sub navigation 1"
-  },
-  items: [{
-    html: '
-      <div class="govuk-grid-row">
-        <div class="govuk-grid-column-three-quarters">
-          <p class="govuk-body govuk-!-font-weight-bold govuk-!-margin-0 govuk-!-padding-0 govuk-!-padding-bottom-1">
-            Your probation practitioner
-          </p>
-        </div>
-        <div class="govuk-grid-column-one-quarter govuk-!-text-align-right">
-          <span class="govuk-body-s govuk-tag govuk-tag--red govuk-box--rounded govuk-!-margin-left-1">Unread</span>
-        </div>
-      </div>
-      <p class="govuk-body-s govuk-!-margin-top-1" >' + message.description + '</p>',
-    attributes: {
-      "aria-label": "Section 1"
-    },
-    classes: "moj-ticket-panel__content--blue"
-  }]
-}) }}
       </div>
 
-      <div class="govuk-grid-row govuk-!-margin-top-2">
-        {{ govukButton({
-          text: "Reply",
-          classes: "",
-          attributes: {
-            id: "reply-button"
-          }
+      <div id="reply-form" class="govuk-grid-row govuk-!-margin-top-6">
+          <div class="govuk-box govuk-box--white govuk-box--rounded govuk-!-padding-6 border box-shadow">
+
+            <div id="message-list">
+        {{ mojMessages({
+          items: message.items
         }) }}
       </div>
 
-      <div id="reply-form" class="govuk-grid-row govuk-!-margin-top-6" style="display: none;">
-          <div class="govuk-box govuk-box--white govuk-box--rounded govuk-!-padding-6 border box-shadow">
+            <form id="replyForm" class="govuk-!-margin-top-2"  method="post" action="/pop/messages/thread/{{message.id}}">
+              <div class="govuk-form-group">
+                <textarea class="govuk-textarea" id="message" name="message" rows="5" placeholder="Enter your message"></textarea>
+              </div>
+              <div class="govuk-grid-row">
+                <div class="govuk-grid-column-full govuk-!-text-align-right">
+                      <input type="hidden" name="_csrf" value="{{ csrfToken }}">
 
-        <form class="">
-          <div class="govuk-form-group">
-            <label class="govuk-label" for="subject">Subject</label>
-            <input class="govuk-input" id="subject" name="subject" type="text" placeholder="">
+                  {{ govukButton({
+                    text: "Cancel",
+                    href: "/pop/messages",
+                    classes: "govuk-button--secondary govuk-!-margin-right-2 govuk-!-margin-left-2"
+                  }) }}
+                  {{ govukButton({
+                    text: "Reply",
+                    type: "submit"
+                  }) }}
+                </div>
+              </div>
+            </form>
           </div>
-          <div class="govuk-form-group">
-            <label class="govuk-label" for="message">Your message</label>
-            <textarea class="govuk-textarea" id="message" name="message" rows="5" placeholder=""></textarea>
-          </div>
-          <div class="govuk-form-group">
-            <label class="govuk-label" for="attachment">Attach supporting evidence (optional)</label>
-            <input class="govuk-file-upload" id="attachment" name="attachment" type="file">
-          </div>
-          <div class="govuk-grid-row">
-            <div class="govuk-grid-column-full govuk-!-text-align-right">
-              {{ govukButton({
-                text: "Cancel",
-                href: "/pop/messages",
-                classes: "govuk-button--secondary govuk-!-margin-right-2 govuk-!-margin-left-2"
-              }) }}
-              {{ govukButton({
-                text: "Reply",
-                type: "submit"
-              }) }}
-            </div>
-          </div>
-        </form>
-        </div>
       </div>
     </div>
   </div>
 </div>
 
-{% endblock %}
-
-{% block bodyEnd %}
-  {{ super() }}
-
-  <script nonce="{{ cspNonce }}">
-  document.addEventListener('DOMContentLoaded', function() {
-    var replyButton = document.getElementById('reply-button');
-    var replyForm = document.getElementById('reply-form');
-
-    if (replyButton && replyForm) { 
-      replyButton.addEventListener('click', function() {
-        replyForm.style.display = 'block';
-      });
-    }
-  });
-  </script>
 {% endblock %}
