+++ types/@ministryofjustice/frontend/moj/filters/all.d.ts
@@ -1,0 +1,4 @@
+declare module '@ministryofjustice/frontend/moj/filters/all.js' {
+  const mojFilters: () => Record<string, (...args: unknown[]) => unknown>
+  export = mojFilters
+}
